Split is now identical to previous output using new eplode#split/join fxn
