+++ lib/explode.js
@@ -123,22 +123,20 @@ class Explode {
      */
     split(streets, addresses) {
         //Break into 1km segments
-        if (!this.settings.noDistance) {
-            let distRes = {
-                type: 'FeatureCollection',
-                features: []
-            }
-
-            for (let feat of results.features) {
-                let dist = turf.lineDistance(feat, 'kilometers');
-
-                for (let dist_it = 0; dist_it < dist; dist_it++) {
-                    distRes.features.push(turf.lineSliceAlong(feat, dist_it, dist_it+1, 'kilometers'));
-                }
-            }
-
-            results = distRes;
+        let distRes = {
+            type: 'FeatureCollection',
+            features: []
         }
+
+        for (let feat of streets.features) {
+            let dist = turf.lineDistance(feat, 'kilometers');
+
+            for (let dist_it = 0; dist_it < dist; dist_it++) {
+                distRes.features.push(turf.lineSliceAlong(feat, dist_it, dist_it+1, 'kilometers'));
+            }
+        }
+
+        return distRes;
     }
 
     /**
