+++ lib/split.js
@@ -143,10 +143,12 @@ function split(nid, cb) {
             coords.push(addr);
         });
 
-        network = explode.line({
+        network = explode.join({
             type: 'FeatureCollection',
             features: [ turf.feature(network) ]
         });
+
+        network = explode.split(network);
 
         let addressCluster = [];
         let numberCluster = [];
